requirements.txt
Python 3.12.2
flask==3.1.1
SQLAlchemy==2.0.41
python-dotenv==1.1.1
flask-cors==6.0.1
flasgger==0.9.7.1
folium==0.20.0
joblib==1.5.1
python-dateutil==2.9.0.post0
referencing==0.36.2
requests==2.32.4
scikit-learn==1.3.2
python-dotenv==1.1.1
psycopg2-binary==2.9.10
Jinja2==3.1.6